requirements.txt
scaleway-sdk
